Add decision problem (#1)
* FEAT: Add class for polygonal curve

* FEAT: Add basic class for free space

* FEAT: Add class for free space

* UPDATE: Remove dummy class

* FEAT: Add a class for computing critical values

* FEAT: Set new epsilon and recompute free space

* UPDATE: Change the location of setEpsilon function readability

* UPDATE: Call computeFreeSpace implicitly in constructor

* FEAT: Add basic class for decision problem

This is a basic form of the class DecisionProblem. If the function checkIfMontonCurveExists which checks if there exists a monotone curve that meets the condition is implemented, this class is completed.

* FEAT: Add function to check if a monotoncurve exists

When two polygonal curves P, Q and an epsilon value is given, the function check if there is a monotone curve that penetrates the free space defined by the polygonal curves and the epsilon value. It uses the method of DFS and traveises the points of L and B which are the member variables of the class FreeSpace in pre-determined order.