FEAT: Add Functions for Approximate GED Calculation with O(sqrt(n)) Approximation Ratio (#3)
* Initial commit

* FEAT: Computes square root n approximation of GED given two polygonal curves

Implemented the (skeleton) function 'computeSquareRootApproxGED' which computes approximation of GED. Other functions, which are used in the function 'computeSquareRootApproxGED', are defined and not implemented, yet. This function follows the 'Algorithm 1' from the refrenced paper.

* UPDATE: Add the type 'CurveAlphabet' and redefine the type 'CurveString'

Redefined the type 'CurveString' by introducing the new type 'CurveAlphabet' that represents an alphabet for transformed curve.

* FEAT: Add copy constructor for the class PolygonalCurve

Added copy constructor for the class PolygonalCurve to use it for the function 'GED::transformCurvesToStrings'.

* FEAT: Add the function 'shiftOrigin' that shifts the grid

This function shifts the grid on which the points lies. It gets a new origin coordinate as an input an shifts all the coordinates of the points of the curve.

* FEAT: Add a function that scales the grid of the curve

Function 'scaleGrid' scales the grid with the scaling factor on which the points of the curve lie. This function devides the coordinate of the points of the curve with the scaling factor.

* FEAT: Add the function 'floorCoordinates' to convert the coordinates of the curve to floor integer values

This function is used in the function 'GED::transformCurvesToStrings'.

* FEAT: Add function 'transformCurvesToStrings' that transforms the polygonal curves into strings

This function gets two polygonal curves and g value as input parametes and return the transformed string pair as an output.

* UPDATE: Update the third parameter of SED

* FEAT: Add the function 'SED'

Function 'SED' checks if GED of given two string has GED less than or equal to given threshold. The algorithm of the function is from Algorithm 3 of referenced paper.

* FEAT: Modify the inequality in SED

* UPDATE: Check if result of DP computation is infinity

* FEAT: Add function 'backtrace' to trace back the DP table and retrieve matching

Function 'backtrace' trace back the DP table to construct the matching for approximated result of GED. Function 'computeCost' uses the result to compute the cost for GED.

* FEAT: Add the function 'computeCost'

Function 'computeCost' computes the cost for GED by summing all the distances between matched points and the size of unmatched point set.

* FIX: Modify the function 'SED'

Initialize the element before sliding for trace back after calculating edit distance and modify the index values. And then added a test code to test the implemented functions.

* UPDATE: Give the random seed number for rand function

* UPDATE: Remove 'using namespace std;' from header files

* UPDATE: Print the examples for Frechet distance

* DEBUG: Fix memory access error in 'L' of function 'SED'

When accessing the last diagonals, the code accessed indexes that are out of range of the variable 'L' of 'SED'. Incremented the size of L to accomodate the edge case.

* UPDATE: Remove unnecessary codes